requirements.txt
Flask==2.0.1
Flask-Cors==3.0.6
flask-paginate==0.7.0
Flask-RESTful==0.3.9
Flask-SimpleLDAP==1.4.0
Flask-SQLAlchemy==2.5.1
h11==0.9.0
httptools==0.1.2
idna==2.10
importlib-metadata==4.4.0
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
jsonify==0.5
login==0.0.6
MarkupSafe==2.0.1
multi-key-dict==2.0.3
paramiko==2.7.2
pbr==5.6.0
pycparser==2.20
pycryptodome==3.10.1
pydantic==1.6.1
PyMySQL==1.0.2
PyNaCl==1.4.0
python-jenkins==1.5.0
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.3.1
requests==2.24.0
SQLAlchemy==1.3.17
starlette==0.13.6
urllib3==1.25.10
websockets==8.0.2
Werkzeug==2.0.1
Flask-Excel==0.0.7
python-gitlab
GitPython==3.0.8
